feat(protocol): simplify protocol and remove prover pool from core protocol (#14541)

+++ packages/protocol/contracts/L1/IProofVerifier.sol
@@ -14,9 +14,9 @@ interface IProofVerifier {
     /// @param blockId Unique identifier for the block.
     /// @param blockProofs Raw bytes representing the proof(s).
     /// @param instance Hashed evidence & config data. If set to zero, proof is
-    /// assumed to be from oracle/system prover.
+    /// assumed to be from oracle prover.
     function verifyProofs(
-        uint256 blockId,
+        uint64 blockId,
         bytes calldata blockProofs,
         bytes32 instance
     )
